refactor: update extraId type to Long in DTOs
This commit changes the data type of the `extraId` field from `Int` to `Long` across various Data Transfer Objects (DTOs). This update affects both request and response DTOs for reports and Berita Acara Pemeriksaan (BAP) related to multiple equipment types.

Affected DTOs:
- `MobileCraneReportRequest` and `MobileCraneReportData` in `MobileCraneReportDto.kt`
- `OverheadCraneBapReportHeader` in `OverheadCraneBapDto.kt`
- `OverheadCraneReportRequest` and `OverheadCraneReportData` in `OverheadCraneReportDto.kt`
- `GantryCraneBapRequest` and `GantryCraneBapReportData` in `GantryCraneBapDto.kt`
- `GantryCraneReportRequest` and `GantryCraneReportData` in `GantryCraneReportDto.kt`
- `ElevatorReportRequest` and `ElevatorReportData` in `ElevatorReportDto.kt`
- `ElevatorBapRequest` and `ElevatorBapReportData` in `ElevatorBapDto.kt`
- `EscalatorReportRequest` and `EscalatorReportData` in `EscalatorReportDto.kt`
- `EscalatorBapRequest` and `EscalatorBapReportData` in `EscalatorBapDto.kt`
- `ForkliftReportRequest` and `ForkliftReportData` in `ForkliftReportDto.kt`
- `GondolaBapRequest` and `GondolaBapReportData` in `GondolaBapDto.kt`
- `ForkliftBapRequest` and `ForkliftBapReportData` in `ForkliftBapDto.kt`
- `MobileCraneBapRequest` and `MobileCraneBapReportData` in `MobileCraneBapDto.kt`
- `GondolaReportRequest` and `GondolaReportData` in `GondolaReportDto.kt`

+++ app/src/main/java/com/nakersolutionid/nakersolutionid/data/remote/dto/escalator/EscalatorReportDto.kt
@@ -6,7 +6,7 @@ import com.nakersolutionid.nakersolutionid.data.remote.dto.common.ResultStatus
 data class EscalatorReportRequest(
     @SerializedName("inspectionType") val inspectionType: String,
     @SerializedName("equipmentType") val equipmentType: String,
-    @SerializedName("extraId") val extraId: Int,
+    @SerializedName("extraId") val extraId: Long,
     @SerializedName("createdAt") val createdAt: String,
     @SerializedName("generalData") val generalData: EscalatorGeneralData,
     @SerializedName("technicalData") val technicalData: EscalatorTechnicalData,
@@ -30,7 +30,7 @@ data class EscalatorReportData(
     @SerializedName("id") val id: String, // From response JSON
     @SerializedName("inspectionType") val inspectionType: String,
     @SerializedName("equipmentType") val equipmentType: String,
-    @SerializedName("extraId") val extraId: Int,
+    @SerializedName("extraId") val extraId: Long,
     @SerializedName("createdAt") val createdAt: String,
     @SerializedName("generalData") val generalData: EscalatorGeneralData,
     @SerializedName("technicalData") val technicalData: EscalatorTechnicalData,
